requirementsCPU.txt
Flask
Flask-Login
Flask-WTF
gunicorn # Docker起動用
Werkzeug
SQLAlchemy
PyMySQL
google-genai
google-api-core
googleapis-common-protos
chromadb
rank-bm25
mecab-python3
pydantic
python-dotenv
openpyxl

bcrypt
requests

ollama
pytest
pytest-mock